ACPI: Interpreter update to fix mutex wait problem This changes the timeout param around the interpreter to a u16, so that ACPI_WAIT_FOREVER is equivalent to 0xFFFF, the value ASL expects to mean "wait forever".

+++ drivers/acpi/events/evxface.c
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
  * Module Name: evxface - External interfaces for ACPI events
- *              $Revision: 131 $
+ *              $Revision: 132 $
  *
  *****************************************************************************/
 
@@ -612,7 +612,7 @@ cleanup:
 
 acpi_status
 acpi_acquire_global_lock (
-	u32                     timeout,
+	u16                     timeout,
 	u32                     *handle)
 {
 	acpi_status             status;
